# perf(app): batch prep