… tests/test_session_queue.py