## Summary

Rework batch prep logic to reduce memory usage and do less work when queuing batches:
- Avoid instantiating heavy pydantic objects
- Reduce serialisation and deserialisation to a minimum
- Mutate in place where we can safely do so
- Add copious comments to the logic
- Update tests to use the revised data structures (no functional changes)

### Profiling

This is single-threaded, CPU-bound operation. Testing on my linux desktop, has an Intel 12600 CPU. The `prepare_values_to_insert` function was measured using the `time` module while running the app in the debugger. Tried with a variety of batch sizes. Threw away the first result, the next 3 iterations were remarkably consistent. Perfect linear scaling too.

#### Before (intolerably slow):
- 100: 0.81s
- 1_000 batch: ~8.1s
- 10_000 batch: ~81s

#### After:
- 100: 0.0042s
- 1_000: 0.042s
- 10_000: 0.42s

Nearly 200x improvement. Even with smaller batch sizes, the time from clicking Invoke button til progress bar moving is noticeably faster.

## Related Issues / Discussions

Offline discussion

## QA Instructions

Try it out. Unit tests for this logic had to be updated to work w/ altered data strctures but the logic in the tests is otherwise the same - everything passing.

I also did some fuzz testing in Canvas and Workflows w/ various batch configs, inspecting created graphs and generations to ensure everything is behaving correctly. Looks good.

## Merge Plan

n/a

## Checklist

- [x] _The PR has a short but descriptive title, suitable for a changelog_
- [ ] _Tests added / updated (if applicable)_
- [ ] _Documentation added / updated (if applicable)_
- [ ] _Updated `What's New` copy (if doing a release after this PR)_