@psychedelicious 2025-02-26
This was cherry-picked into #7687 in the interest of getting a release out this evening.